Use test cases to populate a server.
Creates tools to populate a server and then test that content at a later
time. These will be usefull for migrations from one machine to another
in the case of backing up and restoring, as well as in upgrade jobs.

By creating and testing repositories content in test cases, we can know
if the repositories were created and get clearer output from a CI job
using these tests to populate pulp.